Refactor interactor
